# public/swagger/swagger.json
{
	"components": {
...
		"schemas": {
			"IResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"error": {},
					"data": {},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"error",
					"data",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"Bearer": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT Authorization header using the Bearer scheme."
			}
		}
	},
	"info": {
		"title": "solos_apis",
		"version": "1.0.0",
		"description": "solos apis",
		"license": {
			"name": "license"
		},
		"contact": {
			"name": "BTB"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/transaction/transactionOfBuyFutureRed": {
			"post": {
				"operationId": "TransactionOfBuyFutureRed",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Transaction of buy future red from Admin",
				"tags": [
					"Transaction"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"redToken": {
										"type": "string"
									}
								},
								"required": [
									"redToken"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/transaction/transactionToLeaseProperty": {
			"post": {
				"operationId": "TransactionToLeaseProperty",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Transaction to lease a property",
				"tags": [
					"Transaction"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"agreementId": {
										"type": "string"
									},
									"accountType": {
										"type": "string"
									},
									"propertyId": {
										"type": "string"
									},
									"amount": {
										"type": "string"
									}
								},
								"required": [
									"agreementId",
									"accountType",
									"propertyId",
									"amount"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/transaction/getLeaseTransactionsDetails": {
			"get": {
				"operationId": "GetLeaseTransactionsDetails",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Get transaction Booking/Lease payments of property",
				"tags": [
					"Transaction"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/transaction/getRedAndSolosTransactionByUser": {
			"get": {
				"operationId": "GetRedAndSolosTransactionByUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Transaction"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/transaction/getSolosTransactionByUser": {
			"get": {
				"operationId": "GetSolosTransactionByUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Transaction"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/transaction/saveTransactionOfRedAndSoloReward": {
			"post": {
				"operationId": "SaveTransactionOfRedAndSoloReward",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IResponse"
								}
							}
						}
					}
				},
				"description": "Save transaction of Red and solos reward token",
				"tags": [
					"Transaction"
				],
				"security": [
					{
						"Bearer": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"hashId": {
										"type": "string"
									},
									"walletAddress": {
										"type": "string"
									},
									"transactionType": {
										"type": "string"
									},
									"token": {
										"type": "string"
									},
									"userName": {
										"type": "string"
									}
								},
								"required": [
									"hashId",
									"walletAddress",
									"transactionType",
									"token",
									"userName"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}